+++ packages/ethereum-indexer-db-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ethereum-indexer-db-utils
 
+## 0.6.8
+
+### Patch Changes
+
+- reorg + add streams server (wip)
+- Updated dependencies
+  - ethereum-indexer@0.6.8
+
 ## 0.6.7
 
 ### Patch Changes